Rename /e2e to /integration
