+++ integration/README.md
@@ -1,0 +1,7 @@
+# Integration Tests
+
+This folder contains the integration tests and their utilities of the project.
+
+These tests here are primarily used to test the Addon Operator itself and it's interaction with Kubernetes/OpenShift and OLM.
+
+Previously these tests where named `e2e`-tests, but we choose to rename them, because the real end to end tests are done with osde2e and OCM.
