+++ integration/fixtures_test.go
@@ -1,3 +1,3 @@
-package e2e_test
+package integration_test
 
 var testCatalogSourceImage = "quay.io/operatorhubio/catalog:latest"
