+++ integration/teardown_test.go
@@ -1,4 +1,4 @@
-package e2e_test
+package integration_test
 
 import (
 	"context"
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/openshift/addon-operator/e2e"
+	"github.com/openshift/addon-operator/integration"
 )
 
 func Teardown(t *testing.T) {
@@ -17,7 +17,7 @@ func Teardown(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	objs := e2e.LoadObjectsFromDeploymentFiles(t)
+	objs := integration.LoadObjectsFromDeploymentFiles(t)
 
 	// reverse object order for de-install
 	for i, j := 0, len(objs)-1; i < j; i, j = i+1, j-1 {
@@ -26,7 +26,7 @@ func Teardown(t *testing.T) {
 
 	// Delete all objects to teardown the Addon Operator
 	for _, obj := range objs {
-		err := e2e.Client.Delete(ctx, &obj)
+		err := integration.Client.Delete(ctx, &obj)
 		require.NoError(t, err)
 
 		t.Log("deleted: ", obj.GroupVersionKind().String(),
@@ -37,7 +37,7 @@ func Teardown(t *testing.T) {
 		for _, obj := range objs {
 			// Namespaces can take a long time to be cleaned up and
 			// there is no need to be specific about the object kind here
-			assert.NoError(t, e2e.WaitToBeGone(t, 2*time.Minute, &obj))
+			assert.NoError(t, integration.WaitToBeGone(t, 2*time.Minute, &obj))
 		}
 	})
 }
